feat: StoresProvider implemented + all stores reworked into functions returning Zustand Store
