+++ src/stores/AttractorVisualizer/AttractorVisualizerStatusState.ts
@@ -1,3 +1,4 @@
+/** Zustand store for managing Attractor Visualizer status. */
 export type AttractorVisualizerStatusState = {
   /** Currently selected node state. It should be string consisting of '0' and '1' */
   selectedNodeState: string | null;
