+++ src/stores/AttractorVisualizer/createAttractorVisualizerStatusStore.ts
@@ -1,9 +1,9 @@
 import { create } from 'zustand';
+import type { ZustandStore } from '../ZustandStoreType';
 import type { AttractorVisualizerStatusState } from './AttractorVisualizerStatusState';
 
-/** Zustand store for managing Attractor Visualizer status. */
-const useAttractorVisualizerStatus = create<AttractorVisualizerStatusState>()(
-  (set) => ({
+function createAttractorVisualizerStatusStore(): ZustandStore<AttractorVisualizerStatusState> {
+  return create<AttractorVisualizerStatusState>()((set) => ({
     selectedNodeState: null,
     changeSelectedState: (state) => {
       set({
@@ -14,7 +14,7 @@ const useAttractorVisualizerStatus = create<AttractorVisualizerStatusState>()(
       set({
         selectedNodeState: null,
       }),
-  })
-);
+  }));
+}
 
-export default useAttractorVisualizerStatus;
+export default createAttractorVisualizerStatusStore;
