+++ src/stores/ComputationManager/ComputeEngineStatusStore/ComputeEngineStatusState.ts
@@ -1,5 +1,9 @@
 import type { ComputationStatus } from '../../../types';
 
+/** Zustand store for managing Compute Engine status and computation status.
+ * Provides actions to set and clear the status of the Compute Engine
+ * and the current computation.
+ */
 export type ComputeEngineStatusState = {
   computeEngineStatus: string;
   computationStatus: ComputationStatus;
