+++ src/stores/LiveModel/LoadedModelStore/ModelState.ts
@@ -1,5 +1,8 @@
 import type { ModelType } from '../../../types';
 
+/** Zustand store for loaded model information.
+ *  Manages the ID and type of the currently loaded model in LiveModel.
+ */
 export type ModelState = {
   loadedModelId: number;
   loadedModelType: ModelType;
