+++ src/wiring/GlobalServicesProvider/GlobalServicesProvider.ts
@@ -1,0 +1,9 @@
+class GlobalServicesProvider {
+  //   public computationManagerServ: ComputationManagerInt;
+  //   public liveModelServ: LiveModelInt;
+  //   public tabOperationsServ: TabOperationsInt;
+  //   public resultsOperationsServ: ResultsOperationsInt;
+  //   public warningServ: WarningInt;
+
+  constructor() {}
+}
